managment-system3/graduation_design_init/requirements.txt
pymysql==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.1

